+++ core/src/codecs/default/DefaultCodec.h
@@ -35,6 +35,9 @@ class DefaultCodec : public Codec {
     VectorIndexFormatPtr
     GetVectorIndexFormat() override;
 
+    AttrsIndexFormatPtr
+    GetAttrsIndexFormat() override;
+
     DeletedDocsFormatPtr
     GetDeletedDocsFormat() override;
 
@@ -45,6 +48,7 @@ class DefaultCodec : public Codec {
     VectorsFormatPtr vectors_format_ptr_;
     AttrsFormatPtr attrs_format_ptr_;
     VectorIndexFormatPtr vector_index_format_ptr_;
+    AttrsIndexFormatPtr attrs_index_format_ptr_;
     DeletedDocsFormatPtr deleted_docs_format_ptr_;
     IdBloomFilterFormatPtr id_bloom_filter_format_ptr_;
 };