+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please mark all changes in change log and use the issue from GitHub
 ## Feature
 -   \#2319 Redo metadata to support MVCC
 -   \#2509 Count up query statistics for debug ease
+-   \#2572 Support structured data index
 
 ## Improvement
 -   \#2543 Remove secondary_path related code